LPD-70446 Move to a Fragment Entry Reference generic util to reuse
